Revamp how game history is collected
Instead of iterating over each member of a clan, then iterating over the
list of supported games to create a collection job for each mode, create
a collection job for each member inclusive of all supported modes. The
results are then parsed for supported game modes, reducing the external
call count drastically.

Remove use of redis for caching of member data until a better solution
for the server sizing can be solved.