Set sherpa tracking default disabled
To prevent a situation where a bot is in multiple servers with the same
members, allow for toggling sherpa tracking on/off. This state is likely
to only occur in the bot core server, but it would still be worthwhile
to allow for this feature to be toggled at a user's discrection.