Ensure sherpatime returns valid data
If a member is in a game with one or more sherpas, only the sherpa with
the highest value for time played should count for that game. Prior to
this patch, the _member's_ time was counted irrespective of any and all
sherpa time.

For example, if member_1 spent 3 hours in a game but sherpa_1 spent
2 hours, the command previously would return 3 hours. It now correctly
returns 2.

Fix bug in initial activity history scan on clan sync.