Fix activity tracking
Refactor the pydest execution wrapper to ensure awaitables are not
reused. This also cleans up how connection objects are passed between
functions. Remove the rate limiter as that did not seem to function as
well as expected.

Fix bug in logic for tracking activities for recently active members. If
a member were to log off after doing an activity but before the scan
runs to collect active members and activities, the recently logged off
member would not be properly counted as being in the activity.

Store app config in a class.
Use dictConfig for logging setup.
Add more activity types.
Catch more re-connectable database errors.