Turn the Destiny API response into an object
Cleanup how the response data from the Destiny API is handled. Unless
there is an actual error in connection or auth of the API, pydest will
always return json rather than be fancy with handling private history
or maintenance states. This makes the output more consistent.

Update most references to "Bungie" with "Destiny" to be more consistent.

Allow root log level to be a configurable.