Add clan apply command (#80)
Add new command clan apply to streamline the clan member application
process. This allows a registered user to send an application in the
form of a embed message to a channel that has been designated by server
managers to receive admin notifications.

An admin can then react to this message with either a ✅ or ❌ to
approve or deny the request. An approval will send a clan invite using
the Destiny API and inform the user via DM, denials will only send a DM.

Any pending applications can be recalled using the command clan applied
(as clan pending is used for pending clan invites in the Destiny API)
which then become the canonical messages for reaction based approvals.
Previous messages are deleted to prevent confusion.